+++ packages/eleventy-plugin-mdjs/CHANGELOG.md
@@ -1,6 +1,7 @@
-# @dakmor/eleventy-plugin-mdjs
+# @d4kmor/eleventy-plugin-mdjs
 
 ## 0.0.2
+
 ### Patch Changes
 
 - Initial Release