+++ packages/eleventy-rocket-nav/CHANGELOG.md
@@ -1,6 +1,7 @@
-# @dakmor/eleventy-rocket-nav
+# @d4kmor/eleventy-rocket-nav
 
 ## 0.0.2
+
 ### Patch Changes
 
 - Initial Release