+++ packages/launch/CHANGELOG.md
@@ -1,8 +1,9 @@
-# @dakmor/launch
+# @d4kmor/launch
 
 ## 0.0.2
+
 ### Patch Changes
 
 - Initial Release
 - Updated dependencies [undefined]
-  - @dakmor/cli@0.0.2
+  - @d4kmor/cli@0.0.2