+++ packages/navigation/CHANGELOG.md
@@ -1,6 +1,7 @@
-# @dakmor/navigation
+# @d4kmor/navigation
 
 ## 0.0.2
+
 ### Patch Changes
 
 - Initial Release